IceKirby FSM 구현

+++ QueueEngine_SOURCE/qTime.cpp
@@ -43,7 +43,7 @@ namespace Q
 		wchar_t str[50] = L"";
 		swprintf_s(str, 50, L"Time : %d", (int)fps);
 		
-		int len = wcsnlen_s(str, 50);
+		size_t len = wcsnlen_s(str, 50);
 
 
 		TextOut(hdc, 0, 0, str, len);
